requirements.txt
aiohttp==3.11.11
backoff==2.2.1
beautifulsoup4==4.12.3
colorama==0.4.6
matplotlib==3.9.4
numpy==2.0.2
pandas==2.2.3
Pillow==11.1.0
plotly==5.24.1
python-telegram-bot==21.10
scipy==1.13.1
bs4~=0.0.2
mysql-connector-python~=9.2.0
kaleido==0.2.1